LPS-59420 Move methods doHasTable, hasTable, tableHasColumn and tableHasData from UpgradeProcess to BaseDBProcess
